+++ src/layouts/DropdownMenu.tsx
@@ -1,8 +1,8 @@
 import { forwardRef, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { ReactComponent as LogoutIcon } from "../../../../assets/logout.svg";
-import { AppContext, Unit } from "../../../../context/app-context";
-import Toggle from "../../../Shared/Toggle/Toggle";
+import { ReactComponent as LogoutIcon } from "../assets/logout.svg";
+import { AppContext, Unit } from "../context/app-context";
+import Toggle from "../components/Shared/Toggle/Toggle";
 
 const DropdownMenu = forwardRef<HTMLDivElement>((_, ref) => {
   const { t } = useTranslation();
